# Implement Deepcache Optimization