## Description

[DeepCache](https://github.com/horseee/DeepCache), Yet another optimization

For adjacent timesteps, the result of each layers can be considered 'almost-same' for some cases.

We can just cache them.

Note : this is more beneficial when we have *very many step* such as DDIM environment.

It won't produce dramatic improvement in few-step inference, especially LCM.

The implementation was modified with [gist](https://gist.github.com/mcmonkey4eva/25f2b4d95a5ede0934885d74c0669293) and patched compatibility too.


Speed benchmark with 1.5 models, results will be added:

Vanilla 512x704, 23 step, DPM++ SDE Karras sampler, 2x with Anime6B, 5-sample inference
**2.67 it/s**

Hypertile(All)
![image](https://github.com/AUTOMATIC1111/stable-diffusion-webui/assets/35677394/7bb02c8f-f248-4635-90c0-5fb31b5bfe09)
**3.74 it/s**

DeepCache
![image](https://github.com/AUTOMATIC1111/stable-diffusion-webui/assets/35677394/edf6b163-d6c4-49aa-9928-e66b03bf6a1f)
**3.02 it/s**

DeepCache + HyperTile
**4.59 it/s**

## Compatibility
The optimization is compatible with Controlnet, at least.
**(2.6 it/s, 512x680 2x vs 2.0it/s (without))**
With both, we can achieve 4.7 it/s - yes - it is faster because it reuses whole cache in hires pass.



## Should be tested
We can currently change checkpoint with Refiner / Hires.fix
Then, should be invalidate the cache? or, should we just use it?



## Screenshots/videos:
![image](https://github.com/AUTOMATIC1111/stable-diffusion-webui/assets/35677394/a63f905d-fdad-436c-aafa-04468e8d8212)

Works with Hypertile too, 

## Checklist:

- [ ] I have read [contributing wiki page](https://github.com/AUTOMATIC1111/stable-diffusion-webui/wiki/Contributing)
- [ ] I have performed a self-review of my own code
- [ ] My code follows the [style guidelines](https://github.com/AUTOMATIC1111/stable-diffusion-webui/wiki/Contributing#code-style)
- [ ] My code passes [tests](https://github.com/AUTOMATIC1111/stable-diffusion-webui/wiki/Tests)