@FurkanGozukara 2023-12-05
sounding nice

the hyper tile didnt bring any speed on SDXL

how about this?