@gel-crabs 2023-12-05
> sounding nice
> 
> the hyper tile didnt bring any speed on SDXL
> 
> how about this?

Enormous speed boost. Around 2-3x faster when it kicks in. However, I'm currently unable to get good quality results with it; I think the forward timestep embed patch might need to be adapted to the SGM version.