@aria1th 2023-12-05
@gel-crabs I will do more test within 18 hours, but I guess this should work (as they share same structure)
@FurkanGozukara [The XL Code ](https://github.com/horseee/DeepCache/commit/7c92363413cd93aeac3f22ea30b48eb7a56aff06)has released 5 hours ago, but I will have chance to implement this within day... not immediately. But the code seems to be very large....