@gel-crabs 2023-12-06
> @gel-crabs I will do more test within 18 hours, but I guess this should work (as they share same structure) @FurkanGozukara [The XL Code ](https://github.com/horseee/DeepCache/commit/7c92363413cd93aeac3f22ea30b48eb7a56aff06)has released 5 hours ago, but I will have chance to implement this within day... not immediately. But the code seems to be very large....

I adapted it to use the SGM code and the results are the exact same, so it doesn't need to be further adapted to SGM. I'm gonna do some testing with the in-out blocks and see how it goes.