@aria1th 2023-12-06
Temporary update : I think the implementation should be modified to follow the original paper again.

Original paper says that we should sample the values for nearby steps, not by duration basis.

Although we can only optimize the last final steps, for SD XL, I don't think current one is accurate...  thus this should be fixed again.


![image](https://github.com/AUTOMATIC1111/stable-diffusion-webui/assets/35677394/8d745fac-427d-4d02-be30-f9d4fcc79dee)

![image](https://github.com/AUTOMATIC1111/stable-diffusion-webui/assets/35677394/597ffdd7-c57e-4946-8897-a5c9c625e9a7)

Block indexes : 0, 0, 0
![image](https://github.com/AUTOMATIC1111/stable-diffusion-webui/assets/35677394/7a48b591-fed2-4dfe-bb3f-23324ed7096a)