@gel-crabs 2023-12-06
Alright, I think I've gotten the correct blocks for SDXL:
![Screenshot_20231205_213255](https://github.com/AUTOMATIC1111/stable-diffusion-webui/assets/43506670/6e97de27-3d64-4b88-b011-7944a4ffce2a)
So pretty much just the Cache In Block Indexes changed to 8 and 7.