@aria1th 2023-12-06
768x768 test

Hypertile only
![image](https://github.com/AUTOMATIC1111/stable-diffusion-webui/assets/35677394/65ece4b9-0bbc-4fa2-8adc-21858cf6147a)
7.86it/s


Index 0, 0, 0
![image](https://github.com/AUTOMATIC1111/stable-diffusion-webui/assets/35677394/109baa28-f86a-485b-b546-54909a93dc06)
![image](https://github.com/AUTOMATIC1111/stable-diffusion-webui/assets/35677394/56b8d724-d4dd-474b-a8c7-ba3944be55d1)

Cache Rate 27.23%, 8.03it/s

Index 8, 8, 8
![image](https://github.com/AUTOMATIC1111/stable-diffusion-webui/assets/35677394/78fd2f0f-c43f-445d-8652-b9d2b2a4f3b6)
Cache Rate 27.23%, 8.61it/s

Index 0, 0, 5
![image](https://github.com/AUTOMATIC1111/stable-diffusion-webui/assets/35677394/131cf8fb-8d1b-4a97-957f-363024a1a215)
Cache rate 42.37% 10.8 it/s


0, 0, 6
![image](https://github.com/AUTOMATIC1111/stable-diffusion-webui/assets/35677394/ba8c1cc8-c07b-4b51-9078-8d7d58ea724c)
Cache rate 45.4% 11.1it./s

0, 0, 8
![image](https://github.com/AUTOMATIC1111/stable-diffusion-webui/assets/35677394/407859a9-2ad0-46ed-ab74-f3e37f54d60d)

Cache rate 51.45%, 11.51 it/s



0, 0, 8 + Cache out start timestep 600
![image](https://github.com/AUTOMATIC1111/stable-diffusion-webui/assets/35677394/06ced757-0249-44bf-9729-aeada59ff60a)
46.2%, 10.42 it/s

0, 0, 8 + Cache out start timestep 600 + interval 50
![image](https://github.com/AUTOMATIC1111/stable-diffusion-webui/assets/35677394/69524fa6-a597-44d7-a1d1-410826e02823)
34.9%, 9.18it/s

@gel-crabs 

I think we can use 0, 0, 8 for most case