@FurkanGozukara 2023-12-06
cache looks like degraded quality significantly? @aria1th 

also hyper tile looks like do not degrade quality right?