@gel-crabs 2023-12-06
Oops, didn't see the new update. MUCH less quality loss than before. I'm gonna keep testing and see what I can find.